Changes remaining vector/matrix creation to use non-member functions
The creation of new vectors and matrices, including blocked ones,
is now being performed by non-member functions.  This separates the
object creation from the classes that need and use them.

In the course of making this transition, many unused creation
methods were found in various builders.  These are being removed
for simplification and clarity as to what capabilities are
absolutely required from the linear algebra implementation.

Gitlab issue #36