Add independent CMake for unit tests
Prior to this commit the unit tests could be run only as a part of a
full Xyce build. This commit adds the ability to compile and run the
unit tests outside of a Xyce build. I tried to minimize the number of
files compiled. It works on my system, but I haven't tried it on others.
The README describes how to use the CMake system.

I also tweaked the Xyce-build section for clarity. Also, one test
requires a data file be copied to the test directory. That was added as
well.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/72