Expands Linear::Graph class to include minimum set of required methods
The Linear::Graph class was initially added as a simple wrapper for
an Epetra_CrsGraph object.  Because of that, the epetraObj() method
was necessarily used anytime the object was required to perform an
operation.

This commit adds the minimum number of methods required by Xyce to
the Linear::Graph class and then removes any unnecessary access of
the underlying Epetra_CrsGraph object.  Non-member constructors
for Linear::Graph have been added to the N_LAS_SystemHelpers to
separate the creation of the objects from the class implementing
the objects.

More unnecessary Epetra headers have been removed and, after this
commit, the builders for transient, HB, PCE, and ES are free of
direct interaction with Epetra through the use of non-member
constructors and linear algebra interfaces that can now be abstracted.

Gitlab issue #36