Add many more complex unit tests,from double tests
Issue # 72

"Integrate the new expression library unit tests into CTest"

This commit implements a large number of expression library
parser unit tests, for complex-typed expressions.  These are mostly
tests that I have copied over and modified from the
`parserUnitTest.C` file into the `complexParserUnitTest.C` file.

The development of `complex` parser unit tests has lagged behind the
development of the `double` parser unit tests.  For historical reasons,
I have tended to create tests in the `double` version first.

There are currently around 540 `double` tests, so there
should be *more* than 540 `complex` tests, ideally.

There are a few `complex`-only capabilities that can't be meaningfully
tested using `double`, but there aren't any cases that I can think
of where the opposite is true.  All the `double` tests that I know
of can work as `complex` tests.

So ideally: (number of `complex` tests) > (number of `double` tests)

Prior to this commit there were only approximately 320 `complex`
tests compared to the 540 `double` tests.  With this commit, the number of
`complex` tests is increased to 488.  The work isn't done yet, but this
is a big step in the right direction.

The unit tests are being integrated into the CMake framework. For
various reasons it is difficult to include the `double` tests in
automated regression testing.  The `complex` version of the expression
library is what Xyce actually uses, so it is important to have the
`complex` tests be more complete.