Removing "friend" relationship between linear algebra classes
The use of "friend" is not considered good form, in general.  However,
it is not acceptable in abstact interfaces.  The interfaces should be
well defined enough that classes being "friends" is not necessary.
Turns out that with all the previous work, it was not difficult to
remove this relationship.  One additional method was added to the
Linear::Matrix class to support the Linear::FilteredMatrix class.

A small "nit" was also fixed in the Linear::MultiVector interface
method "daxpy".  It is now called "axpy."  The "d" stands for double
precision and while that is the scalar type for the scaling parameter
still, the vector itself may be a different data type than a double.

Gitlab issue #36