Fix ddxOp::val() function to not core dump w/error
Issue #72.

Today I've been working on the complex unit tests.   Most complex tests
were originally implemented as "double" tests, and then copied over from
the parserUnitTest.C program to the complexParserUnitTest.C program.

As part of my work today, in addition to copying over previously
uncopied tests, I am also revisiting tests that I attempted to get going
a while ago, but failed.   Every so often, I'll copy a test from one
program to the other, but then after modifying it for complex numbers,
it fails, even though it passed just fine when using doubles.

Usually, the test can be fixed by making the "gold" standard use EXACTLY
the same expression that the library would have been using.  Usually
this doesn't matter much, but occasionally the roundoff issues matter,
when trying to get exact matches.

One example of this is when computing derivatives of the pow(a,b)
function.

If the expression is:   ddx(pow(a,b),a)

that means that the expression should return the derivative of pow(a,b)
with respect to a.

I initially set up the gold standard to return:  b*a*a.  But, this
failed, due to roundoff error, even if I used the EXPECT_NEAR macro with
what I thought was a generous tolerance.  Then, I changed the
gold standard to be:  (b/a)*std::pow(a,b).  This exactly matched what
the expression library was doing, and then it passed.

This commit contains a number of fixes like this and un-commenting
previously commented out tests.  The total number of complex tests is
now around 495.

Also included in this commit is a fix for the ddxOp::val() function,
which, when given an invalid input would coredump.   This is relevant to
the ddx8 test, which had been commented out.  The ddxOp::val() function
would catch the invalid input, but would then proceed for a while
afterwards and attempt to access a null pointer.  This was happening b/c
I was using the Xyce::Report::UserError() function to report the error,
and it doesn't immediately stop.