Add unit tests for random operators, from double
Issue #72.

This concludes the push to copy over all relevant type-double unit tests
from parserUnitTest.C and implement them as type-complex unit tests in
the complexParserUnitTest.C program.

There are gaps in the unit test coverage, still, but that will have to
be a different effort.  At this stage, any gap that exists in the
type-complex unit testing also probably exists in the type-double unit
testing (except in the rare cases where a feature is only relevant to
one type and not the other).

As of this commit, there are now 560 type-complex unit tests.  Prior to
this week, there were only 317 type-complex unit tests.