Fix a compiler warning, due to misplaced paren
When I converted the sensitivity code to use the new expression library,
I made a small mistake on one line.  I put one of the right-parens in
the wrong place.  The code still compiled, and all the .SENS tests still
passed, so I didn't notice it.  However, it did cause compiler warnings
to be emitted, depending on the compiler and warning level.  I mostly
did not see these warnings with my clang++ build, but I do see them with
gcc.

Anyway, I finally investigated, and it turned out to be a real mistake.
The line in question was one that would not have mattered to most .SENS
tests.    It was line 342:

It incorrectly said:
if (!objVec[iobj]->expPtr->make_constant(strings[istring], replacement_param.getImmutableValue<double>()),paramType)

And I fixed it to be:
if (!objVec[iobj]->expPtr->make_constant(strings[istring], replacement_param.getImmutableValue<double>(),paramType))

Part of why the first incorrect line compiled is that the make_constant
function has an optional argument for paramType.  So, it can be called
with or without that argument.