Move C++11 configure test to after probe for C++ compiler
When the new autoconf macro to check for C++11 support in the C++
compiler was added, it was added before the check for a C++ compiler.

This worked fine if the user had explicitly set a C++ compiler via the
CXX variable, but would fail if the user was letting configure find
the compiler by itself.

I have moved the call to AX_CXX_COMPILE_STDCXX to be *after* the call
to AC_PROG_CXX.

Now, if the user does NOT specify CXX="somethingspecial" then
configure will look for a C++ compiler using a canonical set of
names (CC, c++, g++, cxx, icpc), and use the first one of these it
finds.

Only then will it check if the compiler in $CXX is C++11 compatible,
and what options are required to make it so.