Fix Sparam, Yparam and Zparam for lower case
I noticed while working on unit tests that I never made any for the RF
parameters, S, Y and Z-params.

So, I started to create one and discovered that the expression library
wouldn't recognize them if they were lower case.  This used to work, but
alas, there was no test to ensure that it never got broken.  (there will
be very soon)

This commit fixes that problem.  When I modified the expression library
for "tablefile", and set it up so that the expression library would
handle IO directly, I had to stop up-casing the entire expression
string.  I had been doing that b/c it was convenient, but up-casing the
whole string doesn't work if that string includes filenames, which need
their case to be preserved on case-dependent file systems.

So, I modified the library so that it would only up-case things that
absolutely needed to be up-cased.  I made a mistake and excluded RF
params from this.  That is fixed with this commit.