Add checks to see if astNodePtr_ is null b4 use
In two places in this file, the astNodePtr_, which is the top level node
of the AST tree was unconditionally being used, without checking first
to see if it was actually a valid pointer.

Earlier in the development of the expression library, this was not a
problem, because if the expression failed to parse, the code would
immediately exit and not attempt to do anything after that.  That
(failure to parse)  was the main use case for how the astNodePtr_
could remain null.

But, this summer I switched the expression library to start using the
error reporting functions in Xyce.  Specifically, I now use the
Xyce::Report::UserError function, which buffers errors so that a bunch
of them can be emitted before exit.  Of course, that means that the
downstream code has to be safer.

So, with this commit, I am making the downstream code a little safer.
I tripped over this issue when attempting to add unit tests for the RF
parameters.

Also, another small change in this commit:  I have been using the
Xyce::Util::fixNan and Xyce::Util::fixInf functions to make sure that
the result returned from newExpression::evaluateFunction was a
reasonable number.

However, I was not applying these functions to the derivatives computed
in newExpression::evaluate.  I was checking for Nan and Inf, but in a
less complete way.  ie, it wasn't set up to properly handle complex
numbers, and I wasn't handling signs correctly, among other things.  The
two utility functions do handle all of those issue correctly, so now
they are applied to the derivatives as they should have been.