Adds abstraction for Linear::Problem class with Epetra inheritance
This linear problem class, Linear::Problem, is the simplest class
to start creating abstract interfaces from.  It is a container for
the matrix/operator and vectors that define a linear problem that
will be solved using one of the many solver approaches provided by
Xyce that are direct or iterative.  The Epetra derived class contains
the Epetra_LinearProblem container that performs the same function,
but it required to pass in data to many of the Epetra-based linear
solvers.

Gitlab issue #36