Partial fix for AC .SENS in parallel
Issue #82.

This commit partially fixes sensitivities for AC in parallel.
Prior to this commit, AC sensitivities would crash or hang in parallel.
Now, they don't do that.   But several of my regression tests
now fail, and fail badly.

Interestingly, they don't all fail.

The fix contained in this commit has to do with the B-vector derivatives,
which are part of the RHS vector for AC sensitivities.  They were not being
communicated properly in parallel, and that caused a hang.

The broken stuff is mostly in the (currently excluded) bsim3
sub-directory.  That suggests that the problem is possibly with the
new code I wrote to compute the numerical J'.  But I won't know until
I dig more.