Moves file writing in solvers to helper functions
This cleans up the interfaces to the solvers by reducing repetitive
blocks of code that are Epetra-specific.  The logic that manages
the pointer to the base linear problem and the transformed linear
problem has been made similar between the solvers.  This is for
ease of understanding and for maintainablity.

Gitlab issue #36