More fixes for AC .SENS in parallel.
Issue #82.

Yesterday, I updated AC sensitivities for parallel, so that they would run
without crashing.  Today, I fixed them a bit more, so that they now get
close to the correct answer.

The main thing that is fixed in this commit is the arithmentic finite
differencing operations that are perfomed to compute J' (derivative of AC
Jacobian w.r.t. p).  The function calls I chose to use appeared to be
supported for block matrices, but in fact they are not, at least in some
cases.  The N_LAS_BlockMatrix is derived from the N_LAS_Matrix, so any
function in the N_LAS_Matrix will appear to be supported in the block matrix.
Anyway, when I did things like dJdp_->add(*ACMatrix_), it compiled fine, and
worked in serial, but was basically a no-op in parallel.   Unsurprisingly,
this resulted in an incorrect J'.

I've replaced the matrix operations that did all these operations on the
global level, and I now do them on the block level instead.  The block
level functions must work, because they are used to contruct the AC Jacobian
to begin with.  So, for example:

dJdp_->block(0, 0).add(ACMatrix_->block(0,0))

works fine in parallel, even if:

dJdp_->add(*ACMatrix_)

only works in serial.  Switching to the block calls results in a
slightly larger number of function calls, but not by much, and it should
be roughly the same number of total operations at this point anyway.
And it appears to mostly work correctly now.