Small cleanup of DC/tran sensitivities
I've been working on refactoring/fixing the AC sensitivity code and
while doing that occasionally browse the DC and Tran sensitivities to
get ideas.

While I was browsing it today, I noticed that there are still some code
remnants from the old-expression method for handling objective
functions.  Since switching to the new expression library, some of this
code isn't needed anymore.  Mainly, the code that computed values that
had to be passed directly into the expression library as an argument.
The new expression library handles this stuff automatically.

This commit removes that stuff.