Fix for AC .SENS expression-objective fn output
Issue #81.

Last night I committed and pushed code to implement AC sensitivities
when using expressions.  All of that code worked, but I had only been
testing it by looking at the output sent to std output.

All sensitivities in Xyce can be sent to std output, if the user
requests.  That was implemented a long time ago to mimic the behavior of
the PSPICE version of sensitivities.  It isn't the main focus of
regression testing, however, and it isn't tested at all for AC
sensitivities.  So, this morning I started looking at creating
regression tests for the new feature, and they didn't work, b/c the
output files had mangled the automatic naming for .SENS, among
a few other smaller problems.

This commit fixes the issue, but I don't plan to push new tests
for a while (until after next develop-master merge).

Here's a more detailed explanation of this commit.  There is a lot of code in
the output manager and various outputters that automatically determines exactly
what outputs need to happen for a .SENS calculation, and automatically sets up
some of the output variable names.  A name like d(obj)/dparam_dir is
mostly created there, given the inputs of "obj" and "param".  For AC
sensitivities, these automated names are different as they have to
indicate real vs imag vs mag vs phase of the objective function.

When I used the "OBJFUNC" parameter (which is what DC and TRAN sensitivities
use) to specify the expression-based objective function for AC sensitivities,
this caused the automated .SENS output naming code to get confused.  It produced
BOTH output names that would have been appropriate for DC sensitivities and
AC sensitivities in the *prn file.  This, of course was wrong, and resulted
in file that wasn't usable.

.SENS output in general need a substantial refactor, and that is NOT what
this commit is.  This commit is a small kludge to undo this DC/AC
sensitivities name mangling.

I've added a new paramter, called ACOBJFUNC, which now serves the same
role as OBJFUNC, but it is what should (for now) be used for AC
sensitivities, if using the expression-based objective functions.   Much
of the automated name expansions performed in the output manager were
based on the contents of OBJFUNC vector.  Now, that vector will be
empty, so the mangling will not happen.

I've also added extra code to make sure that the data structures sent to
the outputters from AC .SENS are all populated properly. That was
another (much easier) problem I had to fix.

With this change, it is possible to create meaningful regression tests
for the new expression-based .AC sensitivities.  But, I plan to wait on
that until after the next code merge.