Fixes const correctness for some parallel and linear algebra objects
Now that the interfaces have been cleaned up and pared down, the
long-standing issue of const correctness can be addressed in these
parts of Xyce.  Communicators are not modified after their creation,
so they should be const, as well as the ParMap objects.  For vectors,
the ParMap and Communicator objects are now const.  For matrices,
the Graph objects, as well as the ParMap and Communicator objects,
are const.  This reduces the number of methods for accessing these
objects from vectors and matrices and also reduces the number of
const_cast calls within the parallel distribution and linear algebra
package.  Most of all, it clarifies the expected usage of these
objects being passed in constructors and returned through accessors.

Gitlab issue #36