Adds new EpetraGraph implementation of abstract Graph base class
This commit splits the previous Linear::Graph class into an abstract
base class and an inherited class for the Epetra implementation.
With the new non-member creation methods, this has minimal impact
on the code.

The MPDE builder had three graphs, based on the idea that the matrix
graph might be split between dFdx, dQdx, and graph that was the
intersection of both.  This isn't used in practice, only one base
graph is used and created, so the builder has been updated to reflect
that.

Gitlab issue #36