Tweak the CMake ADMS plugin project input file
I could not get the CMake-based plugin to work as-is. It complained
about not being able to find Trilinos headers, and when I fixed that I
had linking issues. These changes make everythin work for me (and
others, presumably). There might be things that need to be done to make
it more robust; but it works for now (and others won't have to repeat my
efforts).