Makes Linear::Matrix abstract with implementation in Linear::EpetraMatrix
This commit abstracts the Linear::Matrix and Linear::BlockMatrix classes.
The implementations have been moved into the Linear::EpetraMatrix and
Linear::EpetraBlockMatrix classes, respectively.  The Linear::BlockMatrix
inherits from the Linear::Matrix, so the Linear::EpetraBlockMatrix needs
to provide more of the Linear::Matrix interface than previously.

The impact of this change is minimal due to prior commits that move
the creation of these objects into helper functions and builders.

Gitlab issue #36