Fix indexing mistake in applying .FOUR to .SENS
Issue #119.

An external users reported that when using .FOUR and .SENS in
combination, (ie computing FFTs of sensitivities) that the outputs were
sometimes in the wrong order.

This happened when both traditional ouputs and sensitivities were
requested for the same frequency.  The ordering of the output variables
was then inconsistent with the ordering of the output names.

It turned out that this was very easy to fix.  The output variables
were being inserted into the wrong location in the outputVars_ array -
the index I was using for insertion was too small by one.