Improve CMake's handling of FFT library searches
Since Xyce can use either the Intel MKL FFT capability or FFTW, it needs to be
able to find the right FFT for the build. With this commit, CMake will look for
the MKL FFT capability and determine if Trilinos was built with an awareness of
the MKL, If both are true, then it will use the MKL FFT. Otherwise it will look
for FFTW, and use that if found.

The approach for determining if Tilinos is MKL-aware is rather hackish. The
problem is that Trilinos does not report the MKL as a TPL in the
`Trilinos_TPL_LIST` variable. It does, however, add MKL library files to the
`Trilinos_TPL_LIBRARIES` list. Therefore, in CMake, the
`Trilinos_TPL_LIBRARIES` string is searched for the sub-string, "libmkl". The
intent is to avoid requiring the user to specify all the MKL link lines, etc.,
when that has already been done for Trilinos. It is possible to force the Xyce
CMake to use the Intel FFT, if desired, though.

Issue https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/1