Add "Xyce Plumbing" for Support of .FFT Lines
This commit adds the "Xyce Plumbing" for supporting .FFT lines. It
does NOT provide a functional .FFT capability yet, since the results
have not been extensively validated against HSPICE and/or Matlab.
The technical approach (of using FFTMgr and FFTAnalysis classes) is
closer to the approach used for supporting .MEASURE lines, than to the
approach used to support .FOUR lines.  The remainder of this commit
message will describe what features are (and aren't) supported yet.

All of the qualifiers (START, FROM, STOP, TO, NP, FORMAT, WINDOW,
ALFA, FREQ, FMIN and FMAX) allowed on .FFT lines are parsed now.
Any differences in error handling will be discussed in the Bug
Report.  However, no use is made yet of the values for the FORMAT,
ALFA, FREQ, FMIN and FMAX values.  In addition, only the RECT,
BART, BLACK, HANN and HAMM window functions are implemented.

A .OPTIONS FFT FFT_ACCURATE=<val> capability is implemented for
<val>=0 and <val>=1.  For <val>=0, interpolation is used to determine
the output waveform values at the specified sample times, for each
.FFT line.  For <val>=1, breakpoints are inserted at the specified
sample times, and interpolation is not used.  This is the default
behavior of HSPICE.  The current approach of inserting those
breakpoints (from within the individual FFTAnalysis objects, rather
from a consolidated list in the FFTMgr object) may be inefficient.
The breakpoint tolerance is also set unconditionally in the
constructor for each FFTAnalysis object, rather than being obtained
from StepErrorControl at each accepted time-step.  This is likely
not the best long-term approach for that.  For <val>=0, the current
approach of having a copy of the accepted time-step values stored in
each FFTAnalysis object is inefficient, and should be refactored.

The currently implemented FFT interface may be incorrect, and the
outputter functions (that make the <netlistName>.cir.fft) are
incomplete.  The results for the FFT coefficients generally match
Matlab.  However, the results for the FFT coefficients' magnitude and
phase generally do NOT match HSPICE.  Finaly, .FFT output is only
done to a file, while debug output from each FFTAnalysis object is
dumped (unconditionally) to stdout.

SON Bug 1327
https://joseki-srn.sandia.gov/bugzilla/show_bug.cgi?id=1327