Fix new expression POLY for high-dimensions, and other issues
Issue #123

When I originally implemented the POLY operator in the new expression
library, I only worked on it enough to pass the existing regression test
suite.  The largest dimensions were in the test suite are (1) a 5-input,
1st order found in many Sandler-derived circuits and (2) a lot of 2nd
order expressions.  I don't think there were any for order>2.

So, not suprisingly, my original implementation was broken for order>2.
This commit fixes it.  This should now work for arbitrary order and an
arbitrary number of inputs.  Why anyone would *want* to go to high
dimensions is not clear to me;  A 3rd order, 3-input POLY statement
assumes 40 coefficients, which seems a bit excessive for a manually
specified polynomial.  But I think it works now.

This commit also includes adding parameter support for POLY
coefficients, and also support for .FUNC in POLY statements.  These were
previously not supported, at least in the new expression library.
These issues (.param and .func) are not directly part of issue #123, but
they came up as I attempted to test my fix.

Part of why .param and .func were not supported before this commit is
that to parse POLY statements, I chose to create a different lexer context
for POLY.  This meant that for any types of tokens that might happen
inside of the POLY statement, I needed to create new lexer rules.  I did
a minimal job of this, only supporting raw numbers for coefficients.

This commit also adds a number of unit tests.  There were a number of unit
tests for POLY already, but there were some gaps in testing that this commit
partially addresses.

This adds the following tests:
 -  2nd order, 1-input test, w/params for coefs (Double/Complex_Parser_poly_Test.test10)
 -  2nd order, 1-input test, w/params for coefs and in a .func (Double/Complex_Parser_poly_Test.test10a)
 -  4th order, single input test, w/params (Double/Complex_Parser_poly_Test.test11)
 -  3rd order, 2-input test, w/params (Double/Complex_Parser_poly_Test.test12)
 -  3rd order, 3-input test, w/params (Double/Complex_Parser_poly_Test.test13)
 -  4th order, 4-input test, w/params (Double/Complex_Parser_poly_Test.test14)

All pass for both double and complex.

The first 4 test derivatives.  The last two only test the function
itself, as derivatives get harder to do by hand, when you get up to
341 terms, which is the case for the order=4, N=4 case.

Also, in both files, this commit also combines a bunch of the group
classes into a single class.  Many of these classes evolved as I was
working over the past year, and they were either complete redundant or
nearly redundant.  So, it made sense to get rid of a bunch of them.
In order to avoid an error-prone hand editing of several hundred tests
that previously depended on these now-deleted groups, I made use of
"typedef" to alias the old groups to the new combined group.

Finally, this commit adds a little bit of error checking to POLY, to make sure
that the arguments to POLY are consistent.    My recent refactoring made this
a little more necessary.  Previously, I had relied a bit on parsing rules
to enforce these checks, but my recent changes, to support .param and
.func, made it impossible to do it that way.