Clean-up Outputters for .FOUR and .FFT Lines
This commit cleans up a small nit with the outputters for .FOUR and
.FFT lines.  It is "better form" to invoke basic_ios_all_saver at
the start of their printResult_() functions to save the incoming
state of their output stream (os).  That state is then restored upon
the return from those printResult_() functions.  This change has
no effect in this case, since the precision and format changes
within those functions match the Xyce defaults; but it is "cleaner".

SON Bug 1327
https://joseki-srn.sandia.gov/bugzilla/show_bug.cgi?id=1327