Improvements to .FFT Calculations and Outputter
This commit makes various improvements to the calculations and
outputter for .FFT.  The results now match HSPICE better for simple
tests of the HANN and HAMM window, with the exceptions noted below.

This commit also added partial support for .OPTIONS FFT FFTOUT.
The following metrics are now calculated, and match HSPICE for
simple tests of the HANN and HAMM windows: 1) Total Harmonic
Distortion (THD); 2) Signal to Noise-plus-Distortion Ratio (SNDR);
3) Effective Number Of Bits (ENOB); and 4) Spurious Free Dynamic
Range (SFDR).  These metrics are output if FFTOUT=1 (the default
is 0.)  What is still missing for FFTOUT is the additional output
of "printing out 30 harmonic fundamentals sorted by size".

For a simple test of the HANN and HAMM windows with a 1V DC source
and NP=8, the magnitude, normalized magnitude, THD, SNDR, ENOB and
SFDR results now match HSPICE for the HANN window.  For the HAMM
window, the magnitude, THD, SNDR, ENOB and SFDR results match HSPICE.
The normalized magnitude results do not.  The phase results do not
match HSPICE in either case.

SON Bug 1327
https://joseki-srn.sandia.gov/bugzilla/show_bug.cgi?id=1327