Removes infNormIndex and adds index argument to infNorm
The infNormIndex call is often called after the infNorm call.  This
results in two loops over the vector to get the norm and its index.
It makes more sense to have the index be an optional argument to
the infNorm call, so that if the argument exists the index is returned
too.  This also reduces the vector access to 1 loop over the vector.

This commit also removes minValue and maxValue, as they are no longer
necessary after the nonlinear solver cleanup in commit 49b0c8fbf8.

Gitlab issue #36