More Improvements to .FFT Calculations and Outputter
This commit makes additional fixes and changes to the calculations
and outputter for .FFT lines.  The changes/fixes are:

  1) The magnitude and phase of the DC component are now correct, and
both are output to the <netlistName>.fft file.

  2) The normalized magnitude values now match HSPICE.  They are now
normalized to the value of the largest magnitude, rather than the
magnitude of the first harmonic.

  3) The <netlistName>.fft file now only contains values for the
Normalized Magnitude or the Magnitude, depending on whether the
FORMAT qualifier equals NORM or UNORM.  (NORM is the default value).
The text in that file now makes it clear which format is output.

  4) The concept of a "Noise Floor" has been partially added for
compatibility with HSPICE.  If the magnitude of a harmonic is
<= 1e-10 then the phase of that harmonic is set equal to 0.
This improves testability, but needs further discussion with other
Xyce Team members.

Code for calculating and sorting the "harmonic list" for .OPTIONS
FFT FFTOUT=1 has been added.  However, that list is not output yet.

The .FFT capability is still NOT completely correct.  There are
still issues with the value of the NP-th harmonic, in more complex
netlist with time-variant sources.  In addition, only the HANN and
HAMM windows have been validated.

SON Bug 1327
https://joseki-srn.sandia.gov/bugzilla/show_bug.cgi?id=1327