Removes "linearCombo" method from MultiVector in favor of "update"
The "linearCombo and 5-argument "update" method in the multivector class
are essentially the same except for the scalar at the end of the "update"
method that scales the calling object.  If that scalar is zero, then the
two methods are the identical.  In fact, the implementation with Epetra
uses the same Epetra method for both and just puts a zero as the scalar
for the linearCombo method, where the "update" method passes the scalar.

In an effort to simplify the linear algebra interfaces for porting to
new linear algebra libraries, these two methods can be consolidated.
Methods in the nonlinear solver and time integration packages used the
"linearCombo" method heavily, so those have been updated to use the
"update" method.

Gitlab issue #36