Silence a "reorder" warning
One of the more annoying --- and frequently pointless --- warnings
spewed by C++ compilers is when elements of a class are initialized in
the constructor in an order that differes from the order in which
they're declared in the class.

The compiler automatically reorders initializations so that they
happen in the order that they're declared, but spews a warning when it
does.

There are good reasons for this warning, but it only really matters in
some weird use cases.

The warnings are trivially silenced by remembering to initialize in
the same order that you declare variables in a class.

Though this was only noticed today, it does not actually appear to be
a new warning as the files involved have not changed in ages.