Fix FFTW interface in order to work with even number of samples
The FFTW interface does not work properly when the number of samples
is even. It is reported that the Np/2 harmoinc result was wrong in the
.FFT analysis and Np is a even number.

The FFTW library always generate the correct answers. However, the data
is copied to a wrong location when the number of samples is even in FFTW
interface class. In this case, the Np/2 harmonic result only has the real part.
The real part is incorrectly copied to the imaginary part of the Np/2 harmonic
which should be zero. The IFFT calculation has a similar issue. The real
part of the Np/2 harmonic is overwritten by a wrong data when the number of
samples is even. Both issues are fixed now