Fix ADMS forloop and whileloop bug
An open source user informed us that one of the Verilog-A models
inside a PDK was throwing bizarre compilation errors.  I tracked this
down to that model's use of a "for" loop with what amounted to a
nonlinear condition.

ADMS's original "adms.implicit.xml" templates file has a block of code
that descends into blocks and computes its dependencies.  The block
for forloop and whileloop tree elements had some broken conditionals
that could cause it to recurse into the code block for those
loops *twice*.  It only did so for non-constant conditions.

This had multiple bad effects:
   - the first thing that the "block" code does is to reverse the
     order of items in the block.  That's because on parsing ADMS has
     "pushed" each line's worth of block code into the block data
     structure, so it's backwards.  Calling this twice on a block
     reverses it back to the backwards order.

   - The dependency code pushes information about expressions (notably
     function calls) into the data tree.  Calling it twice results in
     duplicates.

The compilation errors were coming about because Xyce/ADMS was seeing
the duplicate functions in the tree and generating function
precomputation code *twice* instead of the correct one time.  Those
compilation errors were not the worst thing, though, because even
without that error the code in the block was also being emitted
in *reverse* order of how it appeared in the input Verilog.

This commit fixes the problem by making the two calls that
adms.implicit.xml would have done mutually exclusive in our own custom
version of that file.  I am not absolutely sure this is logically
equivalent to what the original author had in mind when he wrote the
original mess, but his reasoning is not something I can query.  All I
know now is that this change breaks NO existing Verilog-A
models (notably BSIM-SOI, which made use of both for and while loops,
but in such a manner that did not trip the bug), and does fix the
incorrect code generation of the problematic model and a toy problem I
constructed to isolate the issue.

I did note that fixing the forloop problem alone was not good enough.
As soon as I did that, BSIM-SOI started getting incorrect code
generated for the while loop.  So I had to fix both in the same way.

Though this commit changes an important file in the Xyce/ADMS
processing chain and that requires that I verify that all existing
code in the DeviceModelPKG/ADMS directory continues to be correct, I
am NOT committing all the freshly regenerated devices using it.  I did
this regeneration but confirmed that nothing changed but their
generation dates.  I therefore am not going to force all Xyce
developers to recompile everything for such a trivial change.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/128