Support NBHARM Qualifier for THD for .MEASURE FFT
This commit adds support for the NBHARM qualifier for the THD
measure type for .MEASURE FFT.  For .FFT lines, the Total Harmonic
Distortion (THD) is calculated using all of the harmonics.  The
NBHARM qualifier allows the THD measure to use only a subset of
the harmonics.  For HSPICE compatibility, the following "special
cases" are supported.  NBHARM=1 returns the value of the noiseFloor_
(which is -200db).  If NBHARM <= 0 or NBHARM > NP/2 (where NP is
the number of points in the FFT), the THD value using all of the
harmonics is returned.  Finally, the THD is always returned in dB,
for .MEASURE FFT.

SON bug 1280
https://joseki-srn.sandia.gov/bugzilla/show_bug.cgi?id=1280